Check for permission when Updating or Removing Pet (#2)

+++ src/main/java/com/ribeiroanibal/adopt/rest/controller/PetController.java
@@ -1,10 +1,12 @@
 package com.ribeiroanibal.adopt.rest.controller;
 
 import com.ribeiroanibal.adopt.AdoptApplication;
+import com.ribeiroanibal.adopt.exception.PermissionException;
 import com.ribeiroanibal.adopt.model.Pet;
 import com.ribeiroanibal.adopt.rest.data.PageWrapper;
 import com.ribeiroanibal.adopt.rest.dto.PetPostDto;
 import com.ribeiroanibal.adopt.rest.dto.PetPutDto;
+import com.ribeiroanibal.adopt.security.PermissionHandler;
 import com.ribeiroanibal.adopt.service.PetService;
 import com.ribeiroanibal.adopt.service.UserService;
 import org.springframework.data.domain.Pageable;
@@ -28,10 +30,14 @@ public class PetController {
 
     private final PetService petService;
     private final UserService userService;
+    private final PermissionHandler permissionHandler;
 
-    public PetController(final PetService petService, final UserService userService) {
+    public PetController(final PetService petService,
+                         final UserService userService,
+                         final PermissionHandler permissionHandler) {
         this.petService = petService;
         this.userService = userService;
+        this.permissionHandler = permissionHandler;
     }
 
     @GetMapping
@@ -55,18 +61,26 @@ public class PetController {
     public ResponseEntity<Pet> updateEntity(@PathVariable("id") final Long id,
                                             @RequestBody @Valid PetPutDto petPutDto) {
         final Pet petToUpdate = petService.getEntity(id);
-        petToUpdate.setName(petPutDto.getName());
-        petToUpdate.setAge(petPutDto.getAge());
-        petToUpdate.setDescription(petPutDto.getDescription());
-        petToUpdate.setPhoto(petPutDto.getPhoto());
-        petToUpdate.setActive(petPutDto.getActive());
-        petToUpdate.setVersion(petPutDto.getVersion());
-        return new ResponseEntity<>(petService.updateEntity(petToUpdate), HttpStatus.OK);
+        if (permissionHandler.canEditOrRemove(petToUpdate.getUser().getId())) {
+            petToUpdate.setName(petPutDto.getName());
+            petToUpdate.setAge(petPutDto.getAge());
+            petToUpdate.setDescription(petPutDto.getDescription());
+            petToUpdate.setPhoto(petPutDto.getPhoto());
+            petToUpdate.setActive(petPutDto.getActive());
+            petToUpdate.setVersion(petPutDto.getVersion());
+            return new ResponseEntity<>(petService.updateEntity(petToUpdate), HttpStatus.OK);
+        } else {
+            throw new PermissionException(Pet.class.getSimpleName(), id);
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteEntity(@PathVariable("id") final Long id) {
-        petService.deleteEntity(id);
-        return ResponseEntity.ok().build();
+        if (permissionHandler.canEditOrRemove(id)) {
+            petService.deleteEntity(id);
+            return ResponseEntity.ok().build();
+        } else {
+            throw new PermissionException(Pet.class.getSimpleName(), id);
+        }
     }
 }
